+++ Examples/Resources/xSQLServerAlwaysOnAvailabilityGroup/1-CreateAvailabilityGroup.ps1
@@ -57,7 +57,6 @@ Configuration Example
             EndPointName = 'HADR'
             Ensure = 'Present'
             Port = 5022
-            AuthorizedUser = 'sa'
             SQLServer = $Node.NodeName
             SQLInstanceName = $Node.SQLInstanceName
             PsDscRunAsCredential = $SysAdminAccount